Add instructions for fixing flatpak under Nvidia container

+++ files/dockerfiles/sidecar-dockerd/start.sh
@@ -1,3 +1,8 @@
 #!/usr/bin/env bash
-rm /var/run/docker.pid
+# Remove the docker.pid file.
+if [ -f /var/run/docker.pid ]; then
+    rm /var/run/docker.pid
+fi
+
+# Run the dockerd daemon.
 exec "$@"
